requirements.txt
python==3.6.10
pytables==3.4.4
numpy==1.19.1
pandas==1.1.1
pyqt==5.9.2
opencv==3.4.2
fire==0.3.1
scikit-learn
pathlib
matplotlib
pip
